requirements.txt
Flask==3.0.0
google-cloud-storage
google-cloud-aiplatform>=1.38.0
